# StringUtilities with xplat-intrinsics (and a little perf-tuning)