## Description

Updated some use of platform specific instrinsics with the xplat-instrinsics, so that more target platforms (like Arm64) can benefit from vectorization.

For the smaller "helper"-methods in the isolated benchmark an improvement is shown, but I don't expect this will show up in more macro-like benchmarks. Main benefit of that change is being more xplat.

For the public entry-point methods (in the internal type :wink:) the switch to xplat-instrinsics and some further optimizations shows quite big improvements.

I don't have a test-setup for Arm64 here, where I can run benchmarks. In theory, due xplat-instrinsics, there should be an improvement, but as said I can't validate this.
Is there a way to run benchmarks for Arm64?

## Results

<details>
  <summary>machine info</summary>

```
BenchmarkDotNet=v0.13.2, OS=Windows 10 (10.0.19043.1889/21H1/May2021Update)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET SDK=7.0.100-rc.2.22426.5
```
</details>

### WidenFourAsciiBytesToUtf16AndWriteToBuffer

```
|  Method |      Mean |     Error |    StdDev | Ratio | RatioSD | Code Size |
|-------- |----------:|----------:|----------:|------:|--------:|----------:|
| Default | 0.7093 ns | 0.0142 ns | 0.0133 ns |  1.00 |    0.00 |      44 B |
|      PR | 0.6203 ns | 0.0028 ns | 0.0025 ns |  0.87 |    0.02 |      41 B |
```

<details>
  <summary>Benchmark code</summary>

Mostly a dump for myself, the actual code is copy & pasted from main-branch and changed.

```c#
using System.Runtime.CompilerServices;
using System.Runtime.InteropServices;
using System.Runtime.Intrinsics;
using System.Runtime.Intrinsics.X86;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

BenchmarkRunner.Run<Bench>();

[DisassemblyDiagnoser]
public unsafe class Bench
{
    private int _value = (byte)'a' | (byte)'b' << 8 | (byte)'c' << 16 | (byte)'d' << 24;
    private char* _output;

    [GlobalSetup]
    public void Setup()
    {
        _output = (char*)NativeMemory.AllocZeroed(100, sizeof(char));
    }

    [GlobalCleanup]
    public void Cleanup() => NativeMemory.Free(_output);

    [Benchmark(Baseline = true, OperationsPerInvoke = 1000)]
    public void A()
    {
        for (int i = 0; i < 1000; ++i)
        {
            WidenFourAsciiBytesToUtf16AndWriteToBuffer(_output, _value);
        }
    }

    [Benchmark(OperationsPerInvoke = 1000)]
    public void B()
    {
        for (int i = 0; i < 1000; ++i)
        {
            WidenFourAsciiBytesToUtf16AndWriteToBuffer1(_output, _value);
        }
    }

    private static void WidenFourAsciiBytesToUtf16AndWriteToBuffer(char* output, int value)
    {
        Vector128<sbyte> vecNarrow = Sse2.ConvertScalarToVector128Int32(value).AsSByte();
        Vector128<ulong> vecWide = Sse2.UnpackLow(vecNarrow, Vector128<sbyte>.Zero).AsUInt64();
        Unsafe.WriteUnaligned(output, Sse2.X64.ConvertToUInt64(vecWide));
    }

    private static void WidenFourAsciiBytesToUtf16AndWriteToBuffer1(char* output, int value)
    {
        Vector128<sbyte> vecNarrow = Vector128.CreateScalarUnsafe(value).AsSByte();
        Vector128<ulong> vecWide = Vector128.Widen(vecNarrow).Lower.AsUInt64();
        Unsafe.WriteUnaligned(output, vecWide.ToScalar());
    }
}
```
</details>

Perf-improvement comes from
```diff
vmovd     xmm0,r9d
-vxorps    xmm1,xmm1,xmm1
-vpunpcklbw xmm0,xmm0,xmm1
+vpmovsxbw xmm0,xmm0
vmovq     r9,xmm0
```
so from a more efficient machine-instruction in this case. Further code-size is smaller (by 3 bytes).

### WidenFourAsciiBytesToUtf16AndCompareToChars

```
|  Method |      Mean |     Error |    StdDev | Ratio | RatioSD | Code Size |
|-------- |----------:|----------:|----------:|------:|--------:|----------:|
| Default | 0.8743 ns | 0.0142 ns | 0.0132 ns |  1.00 |    0.00 |      60 B |
|      PR | 0.7800 ns | 0.0114 ns | 0.0107 ns |  0.89 |    0.02 |      57 B |
```

### WidenTwoAsciiBytesToUtf16AndCompareToChars

```
|  Method |      Mean |     Error |    StdDev | Ratio | RatioSD | Code Size |
|-------- |----------:|----------:|----------:|------:|--------:|----------:|
| Default | 0.8864 ns | 0.0174 ns | 0.0244 ns |  1.00 |    0.00 |      61 B |
|      PR | 0.7780 ns | 0.0096 ns | 0.0089 ns |  0.89 |    0.03 |      58 B |
```

### CheckBytesInAsciiRange

Exact the same machine code generated for Avx2 -> Vector256 and Sse2 -> Vector128.

For the `Vector<T>` the PR produces better machine code:
```diff
vxorps    ymm1,ymm1,ymm1
vpcmpgtb  ymm0,ymm0,ymm1
-vpcmpeqd  ymm1,ymm1,ymm1
-vpcmpeqq  ymm0,ymm0,ymm1
vpmovmskb edx,ymm0
cmp       edx,0FFFFFFFF
```

### TryGetAsciiString

* moved to xplat instrinsics
* dropped Vector<T> code-path (there's no platform kestrel run where neither Vector128 nor Vector256 is acceleareted but Vector<T> is)
* processed the remainder in the vectorized path vectorized too due idempotence
* switched layout to have scalar path (= short inputs) first, and the vectorized path below -- in case of longer inputs the jump down to the vectorized path will be amortized by processing more elemnts at once

In the bench-results the `Length` is chosen in respect to vectorization. 15 is just below vectorization kicks in, 31 is for Vector128 path just before Vector256 kicks in, and 100 is just arbitrary length.

```
|  Method | Length |     Mean |     Error |    StdDev |   Median | Ratio | RatioSD | Code Size |
|-------- |------- |---------:|----------:|----------:|---------:|------:|--------:|----------:|
| Default |     15 | 4.346 ns | 0.1270 ns | 0.2050 ns | 4.357 ns |  1.00 |    0.00 |     446 B |
|      PR |     15 | 4.313 ns | 0.1121 ns | 0.1049 ns | 4.286 ns |  0.97 |    0.05 |     526 B |
|         |        |          |           |           |          |       |         |           |
| Default |     31 | 5.590 ns | 0.1270 ns | 0.1126 ns | 5.581 ns |  1.00 |    0.00 |     446 B |
|      PR |     31 | 3.452 ns | 0.1069 ns | 0.1844 ns | 3.364 ns |  0.64 |    0.04 |     526 B |
|         |        |          |           |           |          |       |         |           |
| Default |    100 | 7.579 ns | 0.1854 ns | 0.3390 ns | 7.582 ns |  1.00 |    0.00 |     446 B |
|      PR |    100 | 6.740 ns | 0.1421 ns | 0.1330 ns | 6.672 ns |  0.92 |    0.05 |     526 B |
```

<details>
  <summary>benchmark code</summary>

Mostly a dump for myself, the actual code is copy & pasted from main-branch and changed.

```c#
using System.Runtime.InteropServices;
using BenchmarkDotNet.Attributes;

Bench bench = new();
bench.Setup();
Console.WriteLine(bench.Default());
string s0 = bench.GetString();
bench.Cleanup();

bench.Setup();
Console.WriteLine(bench.PR());
string s1 = bench.GetString();
bench.Cleanup();
Console.WriteLine(s0);
Console.WriteLine(s1);
Console.WriteLine(s0 == s1);

#if !DEBUG
BenchmarkDotNet.Running.BenchmarkRunner.Run<Bench>();
#endif

[DisassemblyDiagnoser]
[ShortRunJob]
public unsafe class Bench
{
    private byte* _input;
    private char* _output;
    private int _count;

    internal string GetString() => new string(_output);

    [Params(15, 31, 100)]
    public int Length { get; set; } = 32 + 16 + 15;

    [GlobalSetup]
    public void Setup()
    {
        _count = this.Length;
        _input = (byte*)NativeMemory.Alloc((uint)this.Length);
        _output = (char*)NativeMemory.AllocZeroed((uint)this.Length, sizeof(char));

        if (_count < 64)
        {
            int idx = 0;
            for (int i = 'A'; i < 'A' + _count; ++i)
            {
                _input[idx++] = (byte)i;
            }

            _input[0] = (byte)'£';
            _input[_count - 1] = (byte)'£';
        }
        else
        {
            new Span<byte>(_input, _count).Fill((byte)'a');
        }
    }

    [GlobalCleanup]
    public void Cleanup()
    {
        NativeMemory.Free(_input);
        NativeMemory.Free(_output);
    }

    [Benchmark(Baseline = true)]
    public bool Default() => StringUtilities.TryGetLatin1String(_input, _output, _count);

    [Benchmark]
    public bool PR() => StringUtilities1.TryGetLatin1String(_input, _output, _count);
}
```
</details>

The improvement for length 31 is expected (due to processing the remainder vectorized).
The improvement for length 100 I didn't expect, but maybe that's from smaller loop bodies (which is almost always better).

Note: I checked the numbers after just switching to xplat-instrinsics (before further optimizations) and the results were within noise, so on-par.

### TryGetLatin1String

* unified the code with `TryGetAsciiString` by using new C# goodness in form of _static abstract interfaces_

```
|  Method | Length |      Mean |    Error |    StdDev | Ratio | RatioSD | Code Size |
|-------- |------- |----------:|---------:|----------:|------:|--------:|----------:|
| Default |     15 |  7.715 ns | 7.250 ns | 0.3974 ns |  1.00 |    0.00 |     473 B |
|      PR |     15 |  4.315 ns | 1.120 ns | 0.0614 ns |  0.56 |    0.03 |     511 B |
|         |        |           |          |           |       |         |           |
| Default |     31 | 12.317 ns | 9.691 ns | 0.5312 ns |  1.00 |    0.00 |     473 B |
|      PR |     31 |  4.098 ns | 7.102 ns | 0.3893 ns |  0.33 |    0.02 |     511 B |
|         |        |           |          |           |       |         |           |
| Default |    100 |  8.343 ns | 9.522 ns | 0.5219 ns |  1.00 |    0.00 |     473 B |
|      PR |    100 |  8.226 ns | 3.028 ns | 0.1660 ns |  0.99 |    0.08 |     511 B |
```

### BytesOrdinalEqualsStringAndAscii

More or less a re-write with inspiration from existing code :wink:.

```
|  Method | Length |     Mean |     Error |    StdDev | Ratio | RatioSD | Code Size |
|-------- |------- |---------:|----------:|----------:|------:|--------:|----------:|
| Default |     15 | 5.534 ns | 0.1498 ns | 0.1665 ns |  1.00 |    0.00 |     474 B |
|      PR |     15 | 5.234 ns | 0.1425 ns | 0.1399 ns |  0.94 |    0.04 |     653 B |
|         |        |          |           |           |       |         |           |
| Default |     31 | 8.812 ns | 0.1948 ns | 0.2793 ns |  1.00 |    0.00 |     474 B |
|      PR |     31 | 6.478 ns | 0.1696 ns | 0.3227 ns |  0.75 |    0.05 |     653 B |
|         |        |          |           |           |       |         |           |
| Default |    100 | 9.238 ns | 0.2090 ns | 0.3064 ns |  1.00 |    0.00 |     474 B |
|      PR |    100 | 8.890 ns | 0.1388 ns | 0.1159 ns |  0.99 |    0.03 |     653 B |
```