@gfoidl 2022-09-17
@tannergooding I'd be happy if you could have a look on the use of xplat-intrinsics here. Thanks in advance!