@tannergooding 2022-10-07
Possibly this weekend, but if I don't get to it by Monday then likely not until the week of the 24th.

I'm in the middle of buying a house so have been pretty busy the past few weeks. Closing on the 17th, so spending the next couple packing/moving.

@stephentoub and @adamsitnik have both been heavily involved in porting SIMD code in the BCL to use the xplat APIs so they might be good people to look as well.