@stephentoub 2022-10-07
My hope actually is that this code completely goes away, with the new APIs being added in https://github.com/dotnet/runtime/pull/75012.  Maybe we could focus on making those as efficient as possible and ensuring they're sufficient to then delete this code? (Regardless, thanks for your tireless efforts on this @gfoidl.)