@gfoidl 2022-10-07
> Maybe we could focus on making those as efficient as possible and ensuring they're sufficient to then delete this code?

Next week I'll scan these codes and collect what APIs are missing so that this goal can be achieved.
Without having a look now, just from the top of my head, the use-cases here have some special cases so that the ASCII-APIs (solely) aren't sufficient -- but as said, I need to have a proper look.