@stephentoub 2022-10-07
Thanks. Whether it's these APIs or other APIs, I'd like to get to a point where we have enough general-purpose vectorized helpers in the core libraries in .NET 8 that dotnet/aspnetcore doesn't need to do any additional vectorization beyond using those APIs.  That's my north star, at least, and if there's something that will prevent us from getting there, I'm very interested in understanding what it is.