@gfoidl 2022-10-20
For "I'll scan these codes and collect what APIs are missing so that this goal can be achieved." unfortunately I didn't had time so far, but it's scheduled as next item in my agenda.