@gfoidl 2022-10-24
Is there a (reasonable easy) way to build aspnetcore on top of a custom runtime-build?
I'd like to build runtime based on https://github.com/dotnet/runtime/pull/75012, then use that build within aspnetcore for easier evaluation what's possible, what's missing, etc.