@gfoidl 2022-10-24
Let the 🌟 shine 😃 (/cc: @stephentoub)

How it could look when StringUtilities is based on the runtime building-blocks w/o custom code here: https://github.com/dotnet/aspnetcore/compare/d2a1c23d90d4e69df47616b59b7215b3da8cc9f6...25c56209d2e3e8b9f8922b260aff5d2271f5d021

I didn't measure perf, just made it working and that tests are passing.

### What's missing to achieve this?

#### ASCII

In StringUtilities for ASCII values of the range `(0x00, 0x80)` are considered valid.
`Ascii.ToUtf16` treats the whole ASCII range `[0x00, 0x80)` as valid.

That's why in the PoC (diff above) I've put https://github.com/dotnet/aspnetcore/blob/25c56209d2e3e8b9f8922b260aff5d2271f5d021/src/Shared/ServerInfrastructure/StringUtilities.cs#L94-L98 to check for this and make the tests pass. This isn't ideal, as we now have a $O(2n)$ operation where it could be $O(n)$.

Thus something like 
```diff
namespace System.Buffers.Text
{
    public static class Ascii
    {
        // existing methods
-       public static OperationStatus ToUtf16(ReadOnlySpan<byte> source, Span<char> destination, out int bytesConsumed, out int charsWritten);
+       public static OperationStatus ToUtf16(ReadOnlySpan<byte> source, Span<char> destination, out int bytesConsumed, out int charsWritten, bool treatNullAsInvalid = false);
    }
}
```
is needed.

#### Latin1

I don't know how hot Latin1 is here, but as it's special cased in https://github.com/dotnet/aspnetcore/blob/d3259f92851e4772d3230177be5b71be20d3ff6d/src/Servers/Kestrel/Core/src/Internal/Infrastructure/HttpUtilities.cs#L140-L143 I think it's hot enough to be optimized. Besided that standard `Encoding.Latin1` can't be used solely, as 0x00 is considered invalid.

Thus basically the same as for ASCII above applies, i.e.

```diff
namespace System.Buffers.Text
{
+   public static class Latin1
+   {
+       // other methods similar as Ascii?
+       public static OperationStatus ToUtf16(ReadOnlySpan<byte> source, Span<char> destination, out int bytesConsumed, out int charsWritten, bool treatNullAsInvalid = false);
+   }
}
```

If the type `Latin1` seems too heavy, too niche, whatever, as alternative one could use something like https://github.com/dotnet/aspnetcore/compare/25c56209d2e3e8b9f8922b260aff5d2271f5d021...e3afae2dcf436dfe357b5a961f0184e51681325e where Latin1 bytes are expanded to UTF-16 via Asii.ToUtf16 and if non-ASCII is met, then the remainder is done in scalar way. Though this is a naive approach, which should be perf-tested -- I don't have numbers on how likeley Latin1 inputs with ranges `[0x80, 0xFF]` are, but if they are rare then that (simple) approach should be good enough.

### Strategy

0. Finish https://github.com/dotnet/runtime/pull/75012
1. Adjust `Ascii.ToUtf16` to optionally treat `\0` as invalid
2. Use `Ascii.ToUtf16` here also for Latin1 (replacing this PR)
3. Benchmark
4. (optional) Feedback to runtime
5. (optional) Propose special Latin1.ToUtf16