@adityamandaleeka 2023-01-06
@gfoidl I was quickly glancing through older PRs and interpreted the strategy above to mean that the next step is a new PR.

We could keep this open while the work is being done, but perhaps it would be better to track the remaining work in an issue(s) and continue the discussion there. That's probably the best way to move forward, especially since the next step sounds like an API proposal for excluding `\0`.

Thanks!