@gfoidl 2023-01-09
Created https://github.com/dotnet/aspnetcore/issues/45962 to move forward.
This PR can be closed, as it would only move from hw-intrinsics to xplat-intrinsics, which itself is an improvement but for the .NET 8 timeframe we can do base the StringUtilities on core libraries provided APIs.